# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Church Projection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Song {
  id: number;
  title: string;
  author: string;
  lyrics: string;
  mp3Path?: string | null;
  key?: string | null;
  bpm?: number | null;
  category?: string | null;
  collectionId?: number | null;
  collectionName?: string | null;
  coverImage?: string | null;
  isFavorite: boolean;
  createdAt: string;
}

export interface SongVerse {
  index: number;
  label: string;
  lines: string[];
}

export interface CreateSongBody {
  title: string;
  author: string;
  lyrics: string;
  mp3Path?: string;
  key?: string;
  bpm?: number;
  category?: string;
  collectionId?: number;
  coverImage?: string;
}

export interface UpdateSongBody {
  title?: string;
  author?: string;
  lyrics?: string;
  mp3Path?: string;
  key?: string;
  bpm?: number;
  category?: string;
  collectionId?: number;
  coverImage?: string;
  isFavorite?: boolean;
}

export interface Collection {
  id: number;
  name: string;
  coverImage?: string | null;
  songCount: number;
  createdAt: string;
}

export interface CollectionWithSongs {
  id: number;
  name: string;
  coverImage?: string | null;
  songs: Song[];
  createdAt: string;
}

export interface CreateCollectionBody {
  name: string;
  coverImage?: string;
}

export interface Liturgy {
  id: number;
  title: string;
  serviceDate?: string | null;
  totalDurationMinutes: number;
  itemCount: number;
  createdAt: string;
}

export type LiturgyItemType =
  (typeof LiturgyItemType)[keyof typeof LiturgyItemType];

export const LiturgyItemType = {
  louvor: "louvor",
  oracao: "oracao",
  biblia: "biblia",
  mensagem: "mensagem",
  pregacao: "pregacao",
  anuncio: "anuncio",
  outro: "outro",
} as const;

export interface LiturgyItem {
  id: number;
  liturgyId: number;
  type: LiturgyItemType;
  title: string;
  details?: string | null;
  durationMinutes: number;
  order: number;
  songId?: number | null;
}

export interface LiturgyWithItems {
  id: number;
  title: string;
  serviceDate?: string | null;
  items: LiturgyItem[];
  totalDurationMinutes: number;
  createdAt: string;
}

export interface CreateLiturgyBody {
  title: string;
  serviceDate?: string;
}

export type CreateLiturgyItemBodyType =
  (typeof CreateLiturgyItemBodyType)[keyof typeof CreateLiturgyItemBodyType];

export const CreateLiturgyItemBodyType = {
  louvor: "louvor",
  oracao: "oracao",
  biblia: "biblia",
  mensagem: "mensagem",
  pregacao: "pregacao",
  anuncio: "anuncio",
  outro: "outro",
} as const;

export interface CreateLiturgyItemBody {
  type: CreateLiturgyItemBodyType;
  title: string;
  details?: string;
  durationMinutes: number;
  order: number;
  songId?: number;
}

export interface UpdateLiturgyBody {
  title?: string;
  serviceDate?: string;
  items?: CreateLiturgyItemBody[];
}

export type UpdateLiturgyItemBodyType =
  (typeof UpdateLiturgyItemBodyType)[keyof typeof UpdateLiturgyItemBodyType];

export const UpdateLiturgyItemBodyType = {
  louvor: "louvor",
  oracao: "oracao",
  biblia: "biblia",
  mensagem: "mensagem",
  pregacao: "pregacao",
  anuncio: "anuncio",
  outro: "outro",
} as const;

export interface UpdateLiturgyItemBody {
  type?: UpdateLiturgyItemBodyType;
  title?: string;
  details?: string;
  durationMinutes?: number;
  order?: number;
}

export type BibleBookTestament =
  (typeof BibleBookTestament)[keyof typeof BibleBookTestament];

export const BibleBookTestament = {
  old: "old",
  new: "new",
} as const;

export interface BibleBook {
  id: string;
  name: string;
  testament: BibleBookTestament;
  chapterCount: number;
}

export interface BibleVerse {
  book: string;
  bookName: string;
  chapter: number;
  verse: number;
  text: string;
  reference: string;
}

export type ProjectionStateMode =
  | (typeof ProjectionStateMode)[keyof typeof ProjectionStateMode]
  | null;

export const ProjectionStateMode = {
  blank: "blank",
  song: "song",
  bible: "bible",
  announcement: "announcement",
} as const;

export interface ProjectionState {
  isActive: boolean;
  mode?: ProjectionStateMode;
  currentSongId?: number | null;
  currentSongTitle?: string | null;
  currentVerseIndex?: number | null;
  currentVerseLine?: string | null;
  nextVerseLine?: string | null;
  totalVerses?: number | null;
  bibleVerse?: string | null;
  bibleReference?: string | null;
  announcement?: string | null;
  isAudioPlaying: boolean;
  audioVolume: number;
  currentLiturgyId?: number | null;
}

export type ProjectionCommandAction =
  (typeof ProjectionCommandAction)[keyof typeof ProjectionCommandAction];

export const ProjectionCommandAction = {
  show_song: "show_song",
  next_verse: "next_verse",
  prev_verse: "prev_verse",
  blank: "blank",
  show_bible: "show_bible",
  show_announcement: "show_announcement",
  play_audio: "play_audio",
  pause_audio: "pause_audio",
  stop_audio: "stop_audio",
  set_volume: "set_volume",
  set_liturgy: "set_liturgy",
} as const;

export interface ProjectionCommand {
  action: ProjectionCommandAction;
  songId?: number;
  verseIndex?: number;
  bibleVerse?: string;
  bibleReference?: string;
  announcement?: string;
  volume?: number;
  liturgyId?: number;
}

export interface StatsOverview {
  totalSongs: number;
  totalCollections: number;
  totalLiturgies: number;
  recentSongs: Song[];
  favoriteSongs: Song[];
  topCollections: Collection[];
}

export type ListSongsParams = {
  search?: string;
  collectionId?: number;
  category?: string;
};

export type ListCollectionsParams = {
  search?: string;
};

export type SearchBibleParams = {
  q: string;
  book?: string;
};