# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Church Projection API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /collections:
    get:
      operationId: listCollections
      tags: [collections]
      summary: List all collections
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Collections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
    post:
      operationId: createCollection
      tags: [collections]
      summary: Create a collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollectionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
  /collections/{id}:
    get:
      operationId: getCollection
      tags: [collections]
      summary: Get collection with songs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionWithSongs"
...
  /bible/search:
    get:
      operationId: searchBible
      tags: [bible]
      summary: Search Bible verses
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: book
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Verses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BibleVerse"
  /bible/{book}/{chapter}:
    get:
      operationId: getBibleChapter
      tags: [bible]
      summary: Get a Bible chapter
      parameters:
        - name: book
          in: path
          required: true
          schema:
            type: string
        - name: chapter
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chapter verses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BibleVerse"
...
  /projection/control:
    post:
      operationId: controlProjection
      tags: [projection]
      summary: Send projection command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectionCommand"
      responses:
        "200":
          description: New state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionState"
...
components:
  schemas:
...
    Song:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        lyrics:
          type: string
        mp3Path:
          type: string
          nullable: true
        key:
          type: string
          nullable: true
        bpm:
          type: integer
          nullable: true
        category:
          type: string
          nullable: true
        collectionId:
          type: integer
          nullable: true
        collectionName:
          type: string
          nullable: true
        coverImage:
          type: string
          nullable: true
        isFavorite:
          type: boolean
        createdAt:
          type: string
      required: [id, title, author, lyrics, isFavorite, createdAt]
...
    Collection:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        coverImage:
          type: string
          nullable: true
        songCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, songCount, createdAt]
    CollectionWithSongs:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        coverImage:
          type: string
          nullable: true
        songs:
          type: array
          items:
            $ref: "#/components/schemas/Song"
        createdAt:
          type: string
      required: [id, name, songs, createdAt]
    CreateCollectionBody:
      type: object
      properties:
        name:
          type: string
        coverImage:
          type: string
      required: [name]
...
    BibleVerse:
      type: object
      properties:
        book:
          type: string
        bookName:
          type: string
        chapter:
          type: integer
        verse:
          type: integer
        text:
          type: string
        reference:
          type: string
      required: [book, bookName, chapter, verse, text, reference]
    ProjectionState:
      type: object
      properties:
        isActive:
          type: boolean
        mode:
          type: string
          enum: [blank, song, bible, announcement]
          nullable: true
        currentSongId:
          type: integer
          nullable: true
        currentSongTitle:
          type: string
          nullable: true
        currentVerseIndex:
          type: integer
          nullable: true
        currentVerseLine:
          type: string
          nullable: true
        nextVerseLine:
          type: string
          nullable: true
        totalVerses:
          type: integer
          nullable: true
        bibleVerse:
          type: string
          nullable: true
        bibleReference:
          type: string
          nullable: true
        announcement:
          type: string
          nullable: true
        isAudioPlaying:
          type: boolean
        audioVolume:
          type: number
        currentLiturgyId:
          type: integer
          nullable: true
      required: [isActive, isAudioPlaying, audioVolume]
    ProjectionCommand:
      type: object
      properties:
        action:
          type: string
          enum: [show_song, next_verse, prev_verse, blank, show_bible, show_announcement, play_audio, pause_audio, stop_audio, set_volume, set_liturgy]
        songId:
          type: integer
        verseIndex:
          type: integer
        bibleVerse:
          type: string
        bibleReference:
          type: string
        announcement:
          type: string
        volume:
          type: number
        liturgyId:
          type: integer
      required: [action]
...